Funciona la copia de archivos

+++ UD01A01-sample/UD01A01/src/main/java/es/cipfpbatoi/damb/ad/procesador/rutas/ProcesadorRutas.java
@@ -8,8 +8,6 @@ import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.util.ArrayList;
-import java.util.List;
 
 public class ProcesadorRutas {
 
@@ -31,7 +29,6 @@ public class ProcesadorRutas {
 
 					getDirectorios(directorio.getAbsolutePath());
 					resultadoProceso.append("Directorio " + directorio.getAbsoluteFile() + ". \n");
-					
 				}
 
 			} else {
@@ -43,19 +40,14 @@ public class ProcesadorRutas {
 
 					resultadoProceso.append("El archivo " + archivo.getName() + " se encuentra en " + archivo.getPath()
 							+ " y es de tipo: " + getExtension(archivo.getPath()) + ". \n");
-					
+
 					if (configuracionBusqueda.getCopiarDirectorio() != "") {
+
+						File destinoCopia = new File(configuracionBusqueda.getCopiarDirectorio());
 						
-						File destinoCopia = new File (configuracionBusqueda.getCopiarDirectorio());
-						
-						if (destinoCopia.exists()) {
-							
-							copiarArchivo(archivo.getAbsolutePath(),destinoCopia.getAbsolutePath());
-							
-							System.out.println("Se copia el archivo  " + archivo.getName() + " a " + destinoCopia.getAbsolutePath() + ". \n" );
-						}	
+						copiarArchivo(archivo.getAbsolutePath(), destinoCopia.getAbsolutePath());
 					}
-					
+
 				}
 			}
 		}
@@ -93,22 +85,22 @@ public class ProcesadorRutas {
 	public static File[] getDirectorios(String ruta) {
 
 		File carpeta = new File(ruta);
-			
-			FileFilter listadoCarpetas = new FileFilter() {
-				
-				public boolean accept(File carpeta) {
 
-						if (carpeta.exists() && carpeta.isDirectory()) {
-							System.out.println(carpeta.getAbsoluteFile());
-							getDirectorios(carpeta.getAbsolutePath());
-							
-							return true;
-						}
-						
-						return false;	
+		FileFilter listadoCarpetas = new FileFilter() {
+
+			public boolean accept(File carpeta) {
+
+				if (carpeta.exists() && carpeta.isDirectory()) {
+					System.out.println(carpeta.getAbsoluteFile());
+					getDirectorios(carpeta.getAbsolutePath());
+
+					return true;
 				}
-			};
-		File[] carpetas = carpeta.listFiles(listadoCarpetas) ;
+
+				return false;
+			}
+		};
+		File[] carpetas = carpeta.listFiles(listadoCarpetas);
 		return carpetas;
 //
 //		if (list != null) {
@@ -138,34 +130,46 @@ public class ProcesadorRutas {
 	}
 
 	public static void copiarArchivo(String origen, String destino) {
-		File archivoOrigen = new File (origen);
-		File archivoDestino = new File (destino);
-		
-		String a_destino = "$archivoDestino.getAbsolutePath()/$archivoOrigen.getName()" ;
-		
-		try {
-			
-			InputStream entrada = new BufferedInputStream(new FileInputStream(archivoOrigen.getName()));
-			OutputStream salida = new BufferedOutputStream(new FileOutputStream(a_destino));
-			
-			byte[] buffer = new byte[1024];
-			int i;
-			
-			while ((i = entrada.read(buffer)) > 0) {
-				salida.write(buffer, 0 , i);
-				salida.flush();
+		File archivoOrigen = new File(origen);
+		File archivoDestino = new File(destino);
+
+		if (archivoDestino.exists()) {
+
+			String a_destino = archivoDestino.getAbsolutePath() + "\\" + archivoOrigen.getName();
+
+			try {
+
+				InputStream entrada = new BufferedInputStream(new FileInputStream(archivoOrigen.getName()));
+				OutputStream salida = new BufferedOutputStream(new FileOutputStream(a_destino));
+
+				byte[] buffer = new byte[1024];
+				int i;
+
+				while ((i = entrada.read(buffer)) > 0) {
+					salida.write(buffer, 0, i);
+					salida.flush();
+				}
+
+				entrada.close();
+				salida.close();
+
+			} catch (Exception e) {
+
+				e.printStackTrace();
 			}
-			
-			entrada.close();
-			salida.close();
-			
-		}catch(Exception e){
-			
-			e.printStackTrace();
+
+		} else {
+
+			System.out.println("No existe el directorio: " + archivoDestino.getAbsolutePath() + ". \n");
+
+			new File(archivoDestino.getAbsolutePath()).mkdirs();
+
+			copiarArchivo(archivoOrigen.getAbsolutePath(), archivoDestino.getAbsolutePath());
+
 		}
-		
+
 	}
-	
+
 	public static boolean checkDirectorio(String ruta) {
 
 		File archivo = new File(ruta);
